requirements.txt
flask==3.1.1
flask-cors==6.0.1
tensorflow==2.10.0
pillow==11.3.0
numpy==1.24.2
gunicorn==23.0.0
gdown